# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server of yixianquan",
    "version": "1.0.0",
    "title": "易闲圈",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "127.0.0.1:9999",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create user",
        "operationId": "createUser",
        "produces": [
          "application/json;chartset=uft-8"
        ],
        "consumes": [
          "application/json;chartset=uft-8"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "409": {
            "description": "conflict with other users"
          }
        }
      }
    },
    "/users/{userID}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get user by user id",
        "description": "",
        "operationId": "getUserByID",
        "produces": [
          "application/json;chartset=uft-8"
        ],
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string",
            "description": "format is userId_token"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Invalid id supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "modify user by user id",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string",
            "description": "format is userId_token"
          },
          {
            "name": "oldPassword",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "modify success"
          },
          "401": {
            "description": "password error"
          },
          "404": {
            "description": "user not exits"
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "userType": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "avator": {
          "type": "string"
        },
        "nickName": {
          "type": "string"
        },
        "age": {
          "type": "integer",
          "format": "int64"
        },
        "sex": {
          "type": "integer"
        },
        "grade": {
          "type": "integer"
        },
        "major": {
          "type": "string"
        },
        "mailAddr": {
          "type": "string"
        },
        "phoneNum": {
          "type": "string"
        },
        "creditVal": {
          "type": "integer"
        },
        "balance": {
          "type": "integer"
        },
        "tags": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
  },
...
}